backend_final/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
drf-spectacular>=0.27,<1.0
django-cors-headers>=4.3,<5.0
django-environ>=0.11,<1.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<11.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
